SimilaritySearch/requirements.txt
torchaudio
numpy
soundfile
statistics
pandas
pyarrow
torchaudio
matplotlib
torch
transformers
faiss-cpu
gradio
